server/requirements.txt
certifi==2019.6.16
chardet==3.0.4
colorama==0.4.1
Django==2.2.2
django-cors-headers==3.0.2
django-filter==2.1.0
django-taggit==1.1.0
djangorestframework==3.9.4
djangorestframework-filters==1.0.0.dev0
et-xmlfile==1.0.1
httpie==1.0.2
idna==2.8
jdcal==1.4.1
mysqlclient==1.4.4
openpyxl==2.6.3
Pygments==2.4.2
pytz==2019.1
requests==2.22.0
sqlparse==0.3.0
urllib3==1.25.3
django-simple-history==2.10.0
# geos-3.7.0
# gdal-2.4.0
# proj-5.2.0
# proj-datumgrid-1.8